Add MobiletechAuthenticator
- Handles login with mobile phone number.
- Will be used by users who registered via SMS purchase of subscription
  (Mobiletech).

Phone numbers as stored at separate table (and not in `user_meta`) so
we can force uniqueness of phone number. We don't want to allow one
number to be linked to 2 customers.

`MobiletechAuthenticator` expects phone number in field `mobile_phone`
or `username` field already used by `UsernameAuthenticator`. Reason for
this is compatibility with already existing login forms. In case user
uses phone number instead of email, he will be logged (if we find
match). Password is still required for this type of login.

remp/crm#1144