Adding mobiletech payment gateway, refactoring notification
There were two type of changes: refactoring, new features.

Refactoring:

- `MobiletechNotificationEvent` now allows handlers to set
outbound message when the notification is sent. Emitters
of the notification can use it for further processing.
This is required for payment gateway to properly create
link between the message and payment.
- `NotificationHandler` now does all the validation and
creates outbound message prepared to be sent. This change
is important, as until now, only "already sent" messages
were being stored to the storage. The change is happening
because payment gateway needs the outbound message to exist
before it's asynchronously processed and sent.
- Since `SendHandler` has reference to outbound message now,
it doesn't need to do any kind of validation. It only populates
the request based on the outbound message and provided
message content and sends it. If the sending request was
successful, it updates outbound message with mobiletech ID.

New features:

- Adding `Mobiletech` gateway able to process and confirm
the payments with this gateway.
- Adding `ConfirmPaymentHandler` which attempts to confirm
the payment based on the delivery status of message.

remp/crm#1197